Undo Dockerfile commit
Even though Zenith can be built in an Arch Docker container, the compiled executable cannot be executed on the host if the host distribution is not Arch. I tried different hacks but the best way to provide backwards compatibility would be to compile Zenith and required libraries on some older version of Ubuntu and package them together.