Make Flutter rendering asynchronous
Before this commit, the main thread was waiting for Flutter to finish
rendering via a semaphore. However, sometimes Flutter doesn't want to
render anything and the main thread will wait indefinitely. This results
 in a deadlock.

We have no way of knowing if Flutter will actually produce a frame or
not. This commit makes Flutter rendering independent using an
intermediate framebuffer and the main thread will not wait anymore. If
Flutter doesn't want to render anything, the last frame will be drawn
once again to the screen.