Clip surface hitbox to its visible bounds (geometry)
In other words, clients that draw their own shadows now have the right
hitbox. Before this commit, shadows would block pointer events that
would normally go to the view underneath. This commit fixes that, but
leaves 10 pixels on each size for easier resizing. It is an arbitrary
value, but it seems that chromium and firefox change the cursor icon
when the pointer approaches the geometry box, 10 pixels before entering
the box.