Basic edge resize implementation, needs some work
When resizing a window from the top or from the left, its position
should be changed only if and when the client resizes the surface.
Currently the window changes position instantly even though the client
hasn't resized the window, which causes some jerkiness i.e. the window
texture rapidly bouncing back and forth.