Fix a bug where you couldn't interact with windows sometimes
The pointer up event was sometimes not emitted when the Listener was
inside the window and the pointer quickly exited the surface. This
caused windows to ignore further pointer events because move or resize
mode wasn't being canceled. Now, this event is emitted by a global
Listener covering the whole screen which fixes the issue.