Route mouse button events though the Flutter Engine
Before, mouse button events were dispatched to the Flutter Engine and to
the application in parallel. This caused situations where interacting
with Flutter widgets would also send input events to the application in
focus.

Now, mouse events are only sent to the Flutter Engine. The widget
representing the application view will then capture these events and
notify the platform that the app has received mouse events.