Remove visibility_detector dependency
I think the visibility detection breaks when scrolling extents are
modified. When tasks are being animated from left to right after a
swipe, the first task is not updating its texture anymore because the
visibility detector thinks that it's not visible.

Anyway, it's not a big deal. Now, only the focused task is queried for
new frames, like on Android.