Add clipboard support
Security is omitted because wlroots makes it difficult to obtain the
serial of an event.

Some functions like `wlr_seat_touch_notify_down` return the serial as
expected, but `wlr_seat_touch_notify_up` doesn't. This is because when
the function is called, wlroots chooses to either send an actual up
event, or finish a drag and drop operation if one is in progress.

In the latter case, Wayland doesn't require a serial to be sent. wlroots
chose to design their API using the common denominator, so most
functions don't return a serial.