Activate the lock screen when the power button is pressed
When the authentication succeeds, the lockscreen destroys itself after
a fade animation and reveals what's behind.

Some things that are still missing:
- Multithreaded authentication. If the authentication fails, the
pam_unix PAM module will block for 2 seconds. This also blocks the main
thread which is responsible for managing the Wayland server and
displaying an image on the screen. This means the entire compositor will
hang for 2 seconds.
- Block all touch input while the screen is off (when brightness = 0).
- Stop requesting new frames from the Wayland client in focus when the
lockscreen is active.
- Stop rendering while the screen is off.
- Other authentication methods like password, model, and fingerprint.