Make the Flutter task runner more efficient
The task runner used to be scheduled every millisecond even if no event
was queued. Now, we ask the kernel to notify the compositor via a file
descriptor at exactly the right time using timerfd. The task runner will
no longer check for events when there is none. I noticed that the CPU
usage was cut in approximately half.