Remove intermediary render pass of application textures
Send surface information to Flutter on commit, not on map. Flutter know
knows about surfaces, xdg_surfaces, xdg_toplevels, and xdg_popups,
instead of the old "view" abstraction. This will make integrating
subsurfaces easier.

We call `wlr_buffer_lock` when an xdg_surface is unmapped. This makes
sure the DMA buffer is not released right away, and the texture is still
in memory for the close animation.

Destruction of various resources is not handled properly yet. A lot of
code is still commented.