Do double buffering for Wayland surface buffers
Flutter renders asynchronously, and it needs to have the latest surface
buffer without waiting for the client.

Every surface has a buffer chain associated, where we store the buffer
in use by Flutter, and the newest buffer committed by the client on that
surface. The app can render at any speed and Flutter will always use the
latest buffer to render.

This is a step towards triple buffering. Double buffering still blocks
the main thread.