Asynchronously wait for surface buffers instead of blocking
The goal of this commit is to let Flutter always render an idle buffer
instead of letting the GPU driver implicitly synchronize at Flutter's
render thread and block the compositing.

If the client sends DMA buffers, we extract a sync file descriptor using
DMA_BUF_IOCTL_EXPORT_SYNC_FILE.

If the buffer is shared memory, `wlroots` will automatically create a
texture out of it. This texture might not be ready immediately, so we
insert a fence using `glFence()` and extract its file descriptor using
`eglDupNativeFenceFDANDROID()`.

In both cases, once the file descriptor becomes readable, the event loop
will be notified and the buffer will be committed to the appropriate
surface buffer chain.