Implement the partial repaint callbacks of the Flutter engine
I upgraded Flutter from 3.3.10 to 3.7.0, and it apparently introduced
partial repaint, a way to tell Flutter to only redraw damaged regions.
The swapchain keeps track of damage regions for each frame buffer.

`embedder.h` states that if I don't explicitly implement partial repaint
it will automatically repaint the whole screen. This was however not
true. I had awful ghosting like in the Windows XP days. Instead of
downgrading Flutter, I decided to use the new API and solve the problem.

The documentation also states that multiple damage regions can be
specified, but this is again not true. If I give it multiple regions,
Flutter simply ignores them and the ghosting reappears. For this reason,
I simply combine all damage regions into one.

I have a feeling that the devs gave us a header with the final API, even
though it's not fully implemented in the Flutter engine.