Show the power menu when the user long presses the power button
There are 2 buttons, one to power off the machine, and the other one to
restart it.

When a button is pressed, the corresponding method on the
`org.freedesktop.login1` DBus object is called. A custom polkit rule
had to be added to remove the need for authentication for these actions.
Containerized apps should not be allowed to call these methods anyway.